// src/interfaces/Enums.ts
export enum Parameter {
    O2 = 0,
    CO = 1,
    CO2 = 2,
    NO = 3,
    NO2 = 4,
    NOx = 5,   
    SO2 = 6,    
    Dust = 7,
    Flow = 8
}

// export enum ParameterF {
//     Flow = 0,
//     CO = 1,
//     CO2 = 2,
//     NO = 3,
//     NO2 = 4,
//     NOx = 5,   
//     SO2 = 6,    
//     Dust = 7
// }